# apps/estimate-room-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WebSocket-based room estimation service.",
        "title": "Estimate Room API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticates a local user, creates a browser session, and returns the authenticated user payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with email and password",
                "parameters": [
                    {
                        "description": "Credentials and OAuth continue URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/oauth2/authorize": {
            "get": {
                "description": "Validates the authorization request and either issues an auth code for an authenticated browser session or redirects the user to the frontend login page.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "oauth2"
                ],
                "summary": "OAuth2 authorize",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth client ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registered redirect URI",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Must be code",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space separated scopes",
                        "name": "scopes",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Opaque client state",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "PKCE challenge",
                        "name": "code_challenge",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "PKCE method",
                        "name": "code_challenge_method",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OIDC nonce",
                        "name": "nonce",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to frontend login or client redirect URI",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth2/token": {
            "post": {
                "description": "Exchanges an authorization code or refresh token for access, refresh, and ID tokens.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth2"
                ],
                "summary": "Exchange OAuth2 token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization_code or refresh_token",
                        "name": "grant_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "PKCE verifier",
                        "name": "code_verifier",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "OAuth client ID",
                        "name": "client_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Redirect URI",
                        "name": "redirect_uri",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "refresh_token",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oauth2dto.TokenResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperrors.ErrorItem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "pointer": {
                    "type": "string"
                }
            }
        },
        "apperrors.HttpError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apperrors.ErrorItem"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "authdto.LoginDTO": {
            "type": "object",
            "required": [
                "continue",
                "email",
                "password"
            ],
            "properties": {
                "continue": {
                    "type": "string",
                    "maxLength": 2000
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
...
        "authdto.SessionResponse": {
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/authdto.SessionUserResponse"
                }
            }
        },
        "authdto.SessionUserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                }
            }
        },
...
        "oauth2dto.TokenResponseDTO": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "id_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "description": "\"Bearer\"",
                    "type": "string"
                }
            }
        },
...
    }
}